+++ app/src/main/java/com/udinesfata/expenz/data/datasource/local/CategoryLocalDataSource.kt
@@ -2,6 +2,10 @@ package com.udinesfata.expenz.data.datasource.local
 
 import com.udinesfata.expenz.data.datasource.local.database.CategoryDao
 import com.udinesfata.expenz.data.model.local.CategoryDb
+import com.udinesfata.expenz.data.utils.constant.SYNC_OPERATION_CREATE
+import com.udinesfata.expenz.data.utils.constant.SYNC_OPERATION_DELETE
+import com.udinesfata.expenz.data.utils.constant.SYNC_OPERATION_UPDATE
+import com.udinesfata.expenz.data.model.query.CategoryQuery
 
 class CategoryLocalDataSource(
     private val categoryDao: CategoryDao
@@ -10,15 +14,45 @@ class CategoryLocalDataSource(
         return categoryDao.getCategory(id)
     }
 
-    suspend fun createCategory(category: CategoryDb) {
-        categoryDao.createCategory(category)
+    suspend fun getCategories(query: CategoryQuery): List<CategoryDb> {
+        return categoryDao.getCategories()
     }
 
-    suspend fun updateCategory(category: CategoryDb) {
-        categoryDao.updateCategory(category)
+    suspend fun createCategory(category: CategoryDb, fromLocal: Boolean = false) {
+        categoryDao.createCategory(
+            category.copy(
+                isSynced = !fromLocal,
+                syncOperation = SYNC_OPERATION_CREATE
+            )
+        )
     }
 
-    suspend fun deleteCategory(id: Int) {
-        categoryDao.deleteCategory(id)
+    suspend fun createCategories(categories: List<CategoryDb>) {
+        for (category in categories) {
+            categoryDao.createCategory(category)
+        }
+    }
+
+    suspend fun updateCategory(category: CategoryDb, fromLocal: Boolean = false) {
+        categoryDao.updateCategory(
+            category.copy(
+                isSynced = !fromLocal,
+                syncOperation = SYNC_OPERATION_UPDATE
+            )
+        )
+    }
+
+    suspend fun deleteCategory(id: Int, flagOnly: Boolean = false) {
+        if (!flagOnly) {
+            categoryDao.deleteCategory(id)
+        } else {
+            val category = categoryDao.getCategory(id)
+            categoryDao.updateCategory(
+                category!!.copy(
+                    isSynced = false,
+                    syncOperation = SYNC_OPERATION_DELETE
+                )
+            )
+        }
     }
 }
