+++ app/src/main/java/com/udinesfata/expenz/data/datasource/local/TransactionLocalDataSource.kt
@@ -2,6 +2,10 @@ package com.udinesfata.expenz.data.datasource.local
 
 import com.udinesfata.expenz.data.datasource.local.database.TransactionDao
 import com.udinesfata.expenz.data.model.local.TransactionDb
+import com.udinesfata.expenz.data.model.query.TransactionQuery
+import com.udinesfata.expenz.data.utils.constant.SYNC_OPERATION_CREATE
+import com.udinesfata.expenz.data.utils.constant.SYNC_OPERATION_DELETE
+import com.udinesfata.expenz.data.utils.constant.SYNC_OPERATION_UPDATE
 
 class TransactionLocalDataSource(
     private val transactionDao: TransactionDao
@@ -10,15 +14,45 @@ class TransactionLocalDataSource(
         return transactionDao.getTransaction(id)
     }
 
-    suspend fun createTransaction(transaction: TransactionDb) {
-        transactionDao.createTransaction(transaction)
+    suspend fun getTransactions(query: TransactionQuery): List<TransactionDb> {
+        return transactionDao.getTransactions()
     }
 
-    suspend fun updateTransaction(transaction: TransactionDb) {
-        transactionDao.updateTransaction(transaction)
+    suspend fun createTransaction(transaction: TransactionDb, fromLocal: Boolean = false) {
+        transactionDao.createTransaction(
+            transaction.copy(
+                isSynced = !fromLocal,
+                syncOperation = SYNC_OPERATION_CREATE
+            )
+        )
     }
 
-    suspend fun deleteTransaction(id: Int) {
-        transactionDao.deleteTransaction(id)
+    suspend fun createTransactions(transactions: List<TransactionDb>) {
+        for (transaction in transactions) {
+            transactionDao.createTransaction(transaction)
+        }
+    }
+
+    suspend fun updateTransaction(transaction: TransactionDb, fromLocal: Boolean = false) {
+        transactionDao.updateTransaction(
+            transaction.copy(
+                isSynced = !fromLocal,
+                syncOperation = SYNC_OPERATION_UPDATE
+            )
+        )
+    }
+
+    suspend fun deleteTransaction(id: Int, flagOnly: Boolean = false) {
+        if (!flagOnly) {
+            transactionDao.deleteTransaction(id)
+        } else {
+            val transaction = transactionDao.getTransaction(id)
+            transactionDao.updateTransaction(
+                transaction!!.copy(
+                    isSynced = false,
+                    syncOperation = SYNC_OPERATION_DELETE
+                )
+            )
+        }
     }
 }
