+++ app/src/main/java/com/udinesfata/expenz/data/datasource/local/WalletLocalDataSource.kt
@@ -2,6 +2,10 @@ package com.udinesfata.expenz.data.datasource.local
 
 import com.udinesfata.expenz.data.datasource.local.database.WalletDao
 import com.udinesfata.expenz.data.model.local.WalletDb
+import com.udinesfata.expenz.data.model.query.WalletQuery
+import com.udinesfata.expenz.data.utils.constant.SYNC_OPERATION_CREATE
+import com.udinesfata.expenz.data.utils.constant.SYNC_OPERATION_DELETE
+import com.udinesfata.expenz.data.utils.constant.SYNC_OPERATION_UPDATE
 
 class WalletLocalDataSource(
     private val walletDao: WalletDao
@@ -10,15 +14,45 @@ class WalletLocalDataSource(
         return walletDao.getWallet(id)
     }
 
-    suspend fun createWallet(wallet: WalletDb) {
-        walletDao.createWallet(wallet)
+    suspend fun getWallets(query: WalletQuery): List<WalletDb> {
+        return walletDao.getWallets()
     }
 
-    suspend fun updateWallet(wallet: WalletDb) {
-        walletDao.updateWallet(wallet)
+    suspend fun createWallet(wallet: WalletDb, fromLocal: Boolean = false) {
+        walletDao.createWallet(
+            wallet.copy(
+                isSynced = !fromLocal,
+                syncOperation = SYNC_OPERATION_CREATE
+            )
+        )
     }
 
-    suspend fun deleteWallet(id: Int) {
-        walletDao.deleteWallet(id)
+    suspend fun createWallets(wallets: List<WalletDb>) {
+        for (wallet in wallets) {
+            walletDao.createWallet(wallet)
+        }
+    }
+
+    suspend fun updateWallet(wallet: WalletDb, fromLocal: Boolean = false) {
+        walletDao.updateWallet(
+            wallet.copy(
+                isSynced = !fromLocal,
+                syncOperation = SYNC_OPERATION_UPDATE
+            )
+        )
+    }
+
+    suspend fun deleteWallet(id: Int, flagOnly: Boolean = false) {
+        if (!flagOnly) {
+            walletDao.deleteWallet(id)
+        } else {
+            val wallet = walletDao.getWallet(id)
+            walletDao.updateWallet(
+                wallet!!.copy(
+                    isSynced = false,
+                    syncOperation = SYNC_OPERATION_DELETE
+                )
+            )
+        }
     }
 }
